feat: update profile-settings dialogs
