+++ packages/ui/src/components/alert-dialog.tsx
@@ -94,7 +94,7 @@ const AlertDialogTitle = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <AlertDialogPrimitive.Title
     ref={ref}
-    className={cn('text-foreground-1 text-xl font-semibold tracking-tight', className)}
+    className={cn('text-foreground-1 text-xl font-medium tracking-tight', className)}
     {...props}
   />
 ))
