+++ packages/ui/src/views/profile-settings/components/index.ts
@@ -1,3 +1,3 @@
-export * from './ssh-key-create/ssh-key-create-dialog'
-export * from './token-create/token-create-dialog'
-export * from './token-create/token-success-dialog'
+export * from './profile-settings-keys-create/profile-settings-keys-create-dialog'
+export * from './profile-settings-token-create/profile-settings-token-create-dialog'
+export * from './profile-settings-token-create/profile-settings-token-success-dialog'
